radv: Add context reuse lists.
Seems like there is a bad behavior in the amdgpu kernel driver that
causes dead contexts to stick around in amdgpu_sync objects,
causing amdgpu_sync_fence to take a *lot* of time.

With this change there are less dead contexts around, as they will
not be created again every time a device is created.